+++ migrations/versions/586cce08f52a_migration_to_v1_5.py
@@ -35,7 +35,7 @@ def upgrade():
     sa.Column('description_en', sa.String(length=2000), nullable=True),
     sa.Column('price', sa.Numeric(precision=6, scale=2, asdecimal=False),
               nullable=True),
-    sa.Column('created', sa.DateTime(), nullable=False),
+    sa.Column('visible', sa.Boolean(), nullable=False),
     sa.Column('active', sa.Boolean(), nullable=False),
     sa.PrimaryKeyConstraint('slug')
     )