+++ scripts/update_sub_states.py
@@ -7,7 +7,7 @@ Conçu pour être appelé tous les jours à minuit. Envoie également un mail
 au Rezident l'informant du changement d'état.
 
 Ce script peut uniquement être appelé depuis Flask :
-/home/intrarez/intrarez/env/bin/flask script update_sub_states.py
+cd /home/intrarez/intrarez; ./env/bin/flask script update_sub_states.py
 
 12/2021 Loïc 137
 """
@@ -21,7 +21,8 @@ try:
 except ImportError:
     sys.stderr.write(
         "ERREUR - Ce script peut uniquement être appelé depuis Flask :\n"
-        "/home/intrarez/intrarez/env/bin/flask script update_sub_states.py\n"
+        "cd /home/intrarez/intrarez; "
+        "./env/bin/flask script update_sub_states.py\n"
     )
     sys.exit(1)